requirements.txt
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
boltons==21.0.0
bracex==2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.1.8
click-option-group==0.5.9
colorama==0.4.6
cryptography==46.0.5
ds_store==1.3.2
exceptiongroup==1.2.2
face==26.0.0
Flask==3.1.3
glom==25.12.0
googleapis-common-protos==1.73.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mac_alias==2.2.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.23.3
mdurl==0.1.2
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-requests==0.58b0
opentelemetry-instrumentation-threading==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
outcome==1.3.0.post0
packaging==26.0
peewee==3.19.0
protobuf==6.33.5
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.12.0
PySocks==1.7.1
python-dotenv==1.2.2
python-multipart==0.0.22
python-telegram-bot==22.7
referencing==0.37.0
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
ruamel.yaml==0.19.1
ruamel.yaml.clib==0.2.14
selenium==4.41.0
semantic-version==2.10.0
semgrep==1.155.0
sniffio==1.3.1
sortedcontainers==2.4.0
sse-starlette==3.3.2
starlette==0.52.1
tomli==2.0.2
trio==0.33.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
wcmatch==8.5.2
websocket-client==1.9.0
Werkzeug==3.1.7
wrapt==1.17.3
wsproto==1.3.2
zipp==3.23.0
blinker==1.9.0
click==8.3.1
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.7
alembic==1.18.4
blinker==1.9.0
click==8.3.1
Flask==3.1.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
python-dotenv==1.2.2
SQLAlchemy==2.0.48
typing_extensions==4.15.0
Werkzeug==3.1.7
